thêm báo cáo, paypal

+++ src/main/java/mstm/muasamthongminh/muasamthongminh/modules/reports/dto/RevenueReportResponse.java
@@ -4,10 +4,12 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 public class RevenueReportResponse {
     private String date;
-    private Long totalRevenue;
+    private BigDecimal totalRevenue;
 }
